fixed warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@4310 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/geometry/Contour2D.cpp
@@ -509,7 +509,7 @@ namespace
         std::vector<int> triangulation;
         Triangulate::processPoints(pnts, triangulation);
 
-        for(int i=0; i<triangulation.size();i+=3){
+        for(size_t i=0; i<triangulation.size();i+=3){
             const Vector2D<>& p1 = pnts[ triangulation[i * 3 + 0] ];
             const Vector2D<>& p2 = pnts[ triangulation[i * 3 + 1] ];
             const Vector2D<>& p3 = pnts[ triangulation[i * 3 + 2] ];
